Fix some naming errors
