+++ squid_py/agreements/events/verify_reward_condition.py
@@ -1,0 +1,18 @@
+def verify_reward_tokens(event, agreement_id, did, service_agreement, price, consumer_address,
+                         publisher_account):
+    """
+
+    :param event:
+    :param agreement_id:
+    :param did:
+    :param service_agreement:
+    :param price:
+    :param consumer_address:
+    :param publisher_account:
+    :return:
+    """
+    # :TODO: verify that tokens were transfered to publisher
+    pass
+
+
+verifyRewardTokens = verify_reward_tokens
