+++ tests/keeper/test_did_registry.py
@@ -13,6 +13,7 @@ def test_did_registry_contract():
     assert isinstance(did_registry, DIDRegistry)
 
 
+@e2e_test
 def test_did_registry_get_block_number_updated():
     test_id = secrets.token_hex(32)
     assert did_registry.get_block_number_updated(test_id) == 0
